fix: jestとStorybookのダミーデータを一部共通化
