+++ components/molecules/card/stories/card.stories.js
@@ -1,72 +1,7 @@
 import { storiesOf } from '@storybook/vue';
 import { select } from '@storybook/addon-knobs/vue';
 import CardListItem from '../CardListItem.vue';
-
-const POSTS1 = [
-  {
-    title: 'PHP',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: '1~4ヶ月',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-  },
-  {
-    title: 'Laravel',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: '2~3ヶ月',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-  },
-  {
-    title: 'JavaScript(ES6)',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: '６ヶ月',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-  },
-  {
-    title: 'Vue.js',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: '8ヶ月',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-  },
-];
-
-const POSTS2 = [
-  {
-    title: 'ブログサイト',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: 'Nuxt.js(Vue.js), axios, microCMS, BootStrap4, vue-infinite-loading',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-    link: 'https://takatechblog.netlify.app/',
-  },
-  {
-    title: 'マークダウンエディター',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: 'Vue.js, Electron, BootStrap4',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-    link: 'https://github.com/taka1156/markdown-editor',
-  },
-  {
-    title: 'LineBot',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: 'Express.js, axios, @line/bot-sdk , Vercel(Zeit Now)',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-    link: 'https://github.com/taka1156/line-bot',
-  },
-  {
-    title: 'オーディオプレイヤー',
-    img: { url: 'http://placehold.jp/150x150.png' },
-    content1: 'Vue.js, Electron, JSMediaTags',
-    content2:
-      'この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れています。この文章はダミーです。文字の大きさ、量、字間、行間等を確認するために入れてい',
-    link: 'https://github.com/taka1156/audio-player',
-  },
-];
+import { dummyCards, dummyCardsLink } from '@/testdata/testdata.js';
 
 storiesOf('Molecules/Card/CardListItem', module).add(
   'default',
@@ -78,8 +13,8 @@ storiesOf('Molecules/Card/CardListItem', module).add(
         type: Object,
         default: select(
           'card',
-          { default: POSTS1[3], clickable: POSTS2[3] },
-          POSTS1[3]
+          { default: dummyCards[3], clickable: dummyCardsLink[3] },
+          dummyCards[3]
         ),
       },
     },
