+++ components/molecules/profile/stories/profile.stories.js
@@ -1,13 +1,7 @@
 import { storiesOf } from '@storybook/vue';
 import { array } from '@storybook/addon-knobs/vue';
 import ProfileList from '../ProflieList.vue';
-
-const DUMMEY = [
-  'TypeScriptの学習',
-  'Laravel+Vueでサービスを作る',
-  'AWSなどインフラ周りの学習',
-  'CSS/SASSの学習',
-];
+import { dummyProfile } from '@/testdata/testdata.js';
 
 storiesOf('Molecules/Profile/ProfileList', module).add(
   'default',
@@ -17,7 +11,7 @@ storiesOf('Molecules/Profile/ProfileList', module).add(
     props: {
       arr: {
         type: Array,
-        default: array('arr', DUMMEY),
+        default: array('arr', dummyProfile),
       },
     },
   }),
