+++ components/organisms/card/tests/card.test.js
@@ -1,5 +1,6 @@
 import { mount } from '@vue/test-utils';
 import CardList from '../CardList.vue';
+import { dummyCards, dummyCardsLink } from '@/testdata/testdata.js';
 
 // CardList
 describe('CardList', () => {
@@ -10,34 +11,6 @@ describe('CardList', () => {
       },
     });
   };
-
-  const dummyCards = [
-    {
-      title: 'ダミータイトル',
-      img: { url: 'http://placehold.jp/150x150.png' },
-      content2: '詳細',
-    },
-    {
-      title: 'ダミータイトル',
-      img: { url: 'http://placehold.jp/150x150.png' },
-      content2: '詳細',
-    },
-  ];
-
-  const dummyCardsLink = [
-    {
-      title: 'ダミータイトル',
-      img: { url: 'http://placehold.jp/150x150.png' },
-      content2: '詳細',
-      link: 'https://taka1156.site',
-    },
-    {
-      title: 'ダミータイトル',
-      img: { url: 'http://placehold.jp/150x150.png' },
-      content2: '詳細',
-      link: 'https://github.com/taka1156/nuxt-portoflio',
-    },
-  ];
 
   it('jumpメソッド', () => {
     // https://remarkablemark.org/blog/2018/11/17/mock-window-location/#update-for-jsdom-14
